Class renaming for RIM loaders.
