--- interaction-dsl/com.temenos.interaction.rimdsl.generator/src/main/java/com/temenos/interaction/rimdsl/generator/launcher/RIMResourceStateLoaderPure.java
@@ -1,23 +1,0 @@
-package com.temenos.interaction.rimdsl.generator.launcher;
-
-import com.temenos.interaction.core.hypermedia.ResourceState;
-
-import java.util.List;
-import java.util.Map;
-
-/**
- * @author kwieconkowski
- */
-
-public class RIMResourceStateLoaderPure extends RIMResourceStateLoaderTemplate {
-
-    @Override
-    protected List<ResourceState> getResourceStateList(String rimFileName) {
-        return (List<ResourceState>) translator.translateTo(getResourceFromRimFile(rimFileName), RimResourceTranslatorImpl.RimResourceTranslationTarget.LIST_OF_RESOURCE_STATES_PURE);
-    }
-
-    @Override
-    protected Map<String, String> getPropertiesMap(String rimFileName) {
-        return (Map<String, String>) translator.translateTo(getResourceFromRimFile(rimFileName), RimResourceTranslatorImpl.RimResourceTranslationTarget.MAP_STRING_STRING_OF_METHODS_AND_PATHS_PURE);
-    }
-}
